+++ src/users.rs
@@ -1,3 +1,4 @@
 pub mod handler;
 pub mod controller;
 pub mod schema;
+pub mod models;
