+++ tests/auth.rs
@@ -4,7 +4,7 @@ use actix_web::{
     App,
 };
 use fake::{faker::internet::raw::SafeEmail, locales::EN, Fake};
-use social_web_service::{auth::schema::*, get_connection_pool, models::NewUser};
+use social_web_service::{auth::schema::*, get_connection_pool, users::models::NewUser};
 
 #[actix_web::test]
 async fn tests_user_login() {
