+++ tests/users.rs
@@ -6,7 +6,10 @@ use actix_web::{
     App,
 };
 use fake::{faker::internet::raw::SafeEmail, locales::EN, Fake};
-use social_web_service::{get_connection_pool, models::NewUser, users::schema::ResponseUser};
+use social_web_service::{
+    get_connection_pool,
+    users::{models::NewUser, schema::ResponseUser},
+};
 
 #[actix_web::test]
 async fn test_users_crud() {
